Improve mobile user experience
By making the grid layout responsive users on mobile devices no longer
have to scroll horizontally to see all cookies.

If there is not enough space in the header to fit the title and the name of the logged
in user we now wrap the name of the logged in user into the next line.